+++ packages/react-native-sdk/expo-config-plugin/src/withIosScreenCapture/index.ts
@@ -1,12 +1,15 @@
-import { ConfigPlugin } from '@expo/config-plugins';
+import { ConfigPlugin, withPlugins } from '@expo/config-plugins';
 import { ConfigProps } from '../common/types';
-import withEntitlementsAndFiles from './withEntitlementsAndFiles';
-import withIosBroadcastExtension from './withIosBroadcastExtension';
+import withPlistUpdates from './withPlistUpdates';
+import withFilesMod from './withFilesMod';
+import withTarget from './withTarget';
 
 const withIosScreenCapture: ConfigPlugin<ConfigProps> = (config, props) => {
-  config = withIosBroadcastExtension(config, props);
-  config = withEntitlementsAndFiles(config, props);
-  return config;
+  return withPlugins(config, [
+    () => withFilesMod(config, props),
+    () => withTarget(config, props),
+    () => withPlistUpdates(config, props),
+  ]);
 };
 
 export default withIosScreenCapture;
